# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Boutique Pro - API de gestión para boutique venezolana
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Producto {
  id: number;
  codigo: string;
  nombre: string;
  categoria: string;
  talla: string;
  color: string;
  precioUsd: number;
  costoUsd: number;
  stock: number;
  stockMinimo: number;
  creadoEn: string;
}

export interface ProductoInput {
  codigo: string;
  nombre: string;
  categoria: string;
  talla: string;
  color: string;
  precioUsd: number;
  costoUsd?: number;
  stock: number;
  stockMinimo?: number;
}

export interface ProductoUpdate {
  codigo?: string;
  nombre?: string;
  categoria?: string;
  talla?: string;
  color?: string;
  precioUsd?: number;
  costoUsd?: number;
  stock?: number;
  stockMinimo?: number;
}

export interface VentaItem {
  id: number;
  productoId: number;
  nombreProducto: string;
  codigoProducto?: string;
  talla?: string;
  color?: string;
  cantidad: number;
  precioUnitarioUsd: number;
  subtotalUsd: number;
}

export interface Venta {
  id: number;
  numeroFactura: string;
  totalUsd: number;
  tasaBcv: number;
  totalBs: number;
  estado: string;
  /** @nullable */
  notas?: string | null;
  creadoEn: string;
  items: VentaItem[];
}

export interface VentaItemInput {
  productoId: number;
  cantidad: number;
}

export interface VentaInput {
  notas?: string;
  items: VentaItemInput[];
}

export interface CompraItem {
  id: number;
  productoId: number;
  nombreProducto: string;
  codigoProducto?: string;
  cantidad: number;
  costoUnitarioUsd: number;
  subtotalUsd: number;
}

export interface Compra {
  id: number;
  numeroOrden: string;
  proveedor: string;
  totalUsd: number;
  tasaBcv: number;
  totalBs: number;
  /** @nullable */
  notas?: string | null;
  creadoEn: string;
  items: CompraItem[];
}

export interface CompraItemInput {
  productoId: number;
  cantidad: number;
  costoUnitarioUsd: number;
}

export interface CompraInput {
  proveedor: string;
  notas?: string;
  items: CompraItemInput[];
}

export interface Dashboard {
  ventasHoy: number;
  ventasHistoricas: number;
  valorInventario: number;
  totalProductos: number;
  productosStockBajo: number;
  facturasHoy: number;
  gananciasDia: number;
  gananciasHistoricas: number;
  totalComprasHoy: number;
}

export interface ProductoVendido {
  productoId: number;
  nombre: string;
  categoria: string;
  totalVendido: number;
  totalUsd: number;
}

export interface VentaDia {
  fecha: string;
  totalUsd: number;
  cantidadFacturas: number;
}

export interface Configuracion {
  id: number;
  tasaBcv: number;
  fuente?: string;
  actualizadoEn: string;
}

export interface ConfiguracionUpdate {
  tasaBcv: number;
}

export type ListarProductosParams = {
categoria?: string;
talla?: string;
stockBajo?: boolean;
};

export type ListarVentasParams = {
fecha?: string;
estado?: string;
fechaDesde?: string;
fechaHasta?: string;
};

export type ListarComprasParams = {
fechaDesde?: string;
fechaHasta?: string;
};

export type ObtenerProductosMasVendidosParams = {
limite?: number;
};