# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Boutique Pro - API de gestión para boutique venezolana
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: productos
    description: Gestión de inventario de productos
  - name: ventas
    description: Gestión de ventas y facturas
  - name: compras
    description: Gestión de compras y reposición de stock
  - name: configuracion
    description: Configuración del sistema
  - name: resumen
    description: Resumen y estadísticas del negocio

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── PRODUCTOS ────────────────────────────────────────────
  /productos:
    get:
      operationId: listarProductos
      tags: [productos]
      summary: Listar todos los productos
      parameters:
        - name: categoria
          in: query
          schema:
            type: string
        - name: talla
          in: query
          schema:
            type: string
        - name: stockBajo
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Producto"
    post:
      operationId: crearProducto
      tags: [productos]
      summary: Crear nuevo producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoInput"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"

  /productos/{id}:
    get:
      operationId: obtenerProducto
      tags: [productos]
      summary: Obtener producto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "404":
          description: Producto no encontrado
    patch:
      operationId: actualizarProducto
      tags: [productos]
      summary: Actualizar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoUpdate"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
    delete:
      operationId: eliminarProducto
      tags: [productos]
      summary: Eliminar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Producto eliminado

  # ── VENTAS ────────────────────────────────────────────────
  /ventas:
    get:
      operationId: listarVentas
      tags: [ventas]
      summary: Listar facturas
      parameters:
        - name: fecha
          in: query
          schema:
            type: string
        - name: estado
          in: query
          schema:
            type: string
        - name: fechaDesde
          in: query
          schema:
            type: string
        - name: fechaHasta
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de facturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venta"
    post:
      operationId: crearVenta
      tags: [ventas]
      summary: Crear nueva venta (factura)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VentaInput"
      responses:
        "201":
          description: Venta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venta"

  /ventas/{id}:
    get:
      operationId: obtenerVenta
      tags: [ventas]
      summary: Obtener venta por ID con detalle de artículos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venta con detalle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venta"
        "404":
          description: Venta no encontrada

  /ventas/{id}/cancelar:
    post:
      operationId: cancelarVenta
      tags: [ventas]
      summary: Cancelar venta y restaurar stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venta cancelada y stock restaurado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venta"

  # ── COMPRAS ────────────────────────────────────────────────
  /compras:
    get:
      operationId: listarCompras
      tags: [compras]
      summary: Listar compras / reposiciones de stock
      parameters:
        - name: fechaDesde
          in: query
          schema:
            type: string
        - name: fechaHasta
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de compras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Compra"
    post:
      operationId: crearCompra
      tags: [compras]
      summary: Registrar una compra / reposición de stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompraInput"
      responses:
        "201":
          description: Compra registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"

  /compras/{id}:
    get:
      operationId: obtenerCompra
      tags: [compras]
      summary: Obtener compra por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Compra encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Compra"
        "404":
          description: Compra no encontrada

  # ── RESUMEN ───────────────────────────────────────────────
  /resumen/dashboard:
    get:
      operationId: obtenerDashboard
      tags: [resumen]
      summary: Obtener resumen del negocio (KPIs)
      responses:
        "200":
          description: Resumen del negocio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"

  /resumen/productos-mas-vendidos:
    get:
      operationId: obtenerProductosMasVendidos
      tags: [resumen]
      summary: Obtener los productos más vendidos
      parameters:
        - name: limite
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Productos más vendidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductoVendido"

  /resumen/ventas-por-dia:
    get:
      operationId: obtenerVentasPorDia
      tags: [resumen]
      summary: Obtener ventas agrupadas por día (últimos 30 días)
      responses:
        "200":
          description: Ventas por día
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VentaDia"

  # ── CONFIGURACIÓN ─────────────────────────────────────────
  /configuracion:
    get:
      operationId: obtenerConfiguracion
      tags: [configuracion]
      summary: Obtener configuración del sistema
      responses:
        "200":
          description: Configuración actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configuracion"
    patch:
      operationId: actualizarConfiguracion
      tags: [configuracion]
      summary: Actualizar configuración (tasa BCV, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfiguracionUpdate"
      responses:
        "200":
          description: Configuración actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configuracion"

  /configuracion/actualizar-tasa-bcv:
    post:
      operationId: actualizarTasaBcvDesdeApi
      tags: [configuracion]
      summary: Obtener y guardar tasa BCV desde API externa
      responses:
        "200":
          description: Tasa BCV actualizada desde API externa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configuracion"
        "503":
          description: No se pudo obtener la tasa desde la API externa

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Producto:
      type: object
      required: [id, codigo, nombre, categoria, talla, color, precioUsd, costoUsd, stock, stockMinimo, creadoEn]
      properties:
        id:
          type: integer
        codigo:
          type: string
        nombre:
          type: string
        categoria:
          type: string
        talla:
          type: string
        color:
          type: string
        precioUsd:
          type: number
        costoUsd:
          type: number
        stock:
          type: integer
        stockMinimo:
          type: integer
        creadoEn:
          type: string

    ProductoInput:
      type: object
      required: [codigo, nombre, categoria, talla, color, precioUsd, stock]
      properties:
        codigo:
          type: string
        nombre:
          type: string
        categoria:
          type: string
        talla:
          type: string
        color:
          type: string
        precioUsd:
          type: number
        costoUsd:
          type: number
        stock:
          type: integer
        stockMinimo:
          type: integer

    ProductoUpdate:
      type: object
      properties:
        codigo:
          type: string
        nombre:
          type: string
        categoria:
          type: string
        talla:
          type: string
        color:
          type: string
        precioUsd:
          type: number
        costoUsd:
          type: number
        stock:
          type: integer
        stockMinimo:
          type: integer

    VentaItem:
      type: object
      required: [id, productoId, nombreProducto, cantidad, precioUnitarioUsd, subtotalUsd]
      properties:
        id:
          type: integer
        productoId:
          type: integer
        nombreProducto:
          type: string
        codigoProducto:
          type: string
        talla:
          type: string
        color:
          type: string
        cantidad:
          type: integer
        precioUnitarioUsd:
          type: number
        subtotalUsd:
          type: number

    Venta:
      type: object
      required: [id, numeroFactura, totalUsd, tasaBcv, totalBs, estado, creadoEn, items]
      properties:
        id:
          type: integer
        numeroFactura:
          type: string
        totalUsd:
          type: number
        tasaBcv:
          type: number
        totalBs:
          type: number
        estado:
          type: string
        notas:
          type: ["string", "null"]
        creadoEn:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/VentaItem"

    VentaItemInput:
      type: object
      required: [productoId, cantidad]
      properties:
        productoId:
          type: integer
        cantidad:
          type: integer

    VentaInput:
      type: object
      required: [items]
      properties:
        notas:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/VentaItemInput"

    CompraItem:
      type: object
      required: [id, productoId, nombreProducto, cantidad, costoUnitarioUsd, subtotalUsd]
      properties:
        id:
          type: integer
        productoId:
          type: integer
        nombreProducto:
          type: string
        codigoProducto:
          type: string
        cantidad:
          type: integer
        costoUnitarioUsd:
          type: number
        subtotalUsd:
          type: number

    Compra:
      type: object
      required: [id, numeroOrden, proveedor, totalUsd, tasaBcv, totalBs, creadoEn, items]
      properties:
        id:
          type: integer
        numeroOrden:
          type: string
        proveedor:
          type: string
        totalUsd:
          type: number
        tasaBcv:
          type: number
        totalBs:
          type: number
        notas:
          type: ["string", "null"]
        creadoEn:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CompraItem"

    CompraItemInput:
      type: object
      required: [productoId, cantidad, costoUnitarioUsd]
      properties:
        productoId:
          type: integer
        cantidad:
          type: integer
        costoUnitarioUsd:
          type: number

    CompraInput:
      type: object
      required: [proveedor, items]
      properties:
        proveedor:
          type: string
        notas:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CompraItemInput"

    Dashboard:
      type: object
      required: [ventasHoy, ventasHistoricas, valorInventario, totalProductos, productosStockBajo, facturasHoy, gananciasDia, gananciasHistoricas, totalComprasHoy]
      properties:
        ventasHoy:
          type: number
        ventasHistoricas:
          type: number
        valorInventario:
          type: number
        totalProductos:
          type: integer
        productosStockBajo:
          type: integer
        facturasHoy:
          type: integer
        gananciasDia:
          type: number
        gananciasHistoricas:
          type: number
        totalComprasHoy:
          type: number

    ProductoVendido:
      type: object
      required: [productoId, nombre, categoria, totalVendido, totalUsd]
      properties:
        productoId:
          type: integer
        nombre:
          type: string
        categoria:
          type: string
        totalVendido:
          type: integer
        totalUsd:
          type: number

    VentaDia:
      type: object
      required: [fecha, totalUsd, cantidadFacturas]
      properties:
        fecha:
          type: string
        totalUsd:
          type: number
        cantidadFacturas:
          type: integer

    Configuracion:
      type: object
      required: [id, tasaBcv, actualizadoEn]
      properties:
        id:
          type: integer
        tasaBcv:
          type: number
        fuente:
          type: string
        actualizadoEn:
          type: string

    ConfiguracionUpdate:
      type: object
      required: [tasaBcv]
      properties:
        tasaBcv:
          type: number