# lib/api-zod/src/generated/types/compra.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Boutique Pro - API de gestión para boutique venezolana
 * OpenAPI spec version: 0.1.0
 */
import type { CompraItem } from './compraItem';

export interface Compra {
  id: number;
  numeroOrden: string;
  proveedor: string;
  totalUsd: number;
  tasaBcv: number;
  totalBs: number;
  /** @nullable */
  notas?: string | null;
  creadoEn: string;
  items: CompraItem[];
}