# lib/api-zod/src/generated/types/compraInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Boutique Pro - API de gestión para boutique venezolana
 * OpenAPI spec version: 0.1.0
 */
import type { CompraItemInput } from './compraItemInput';

export interface CompraInput {
  proveedor: string;
  notas?: string;
  items: CompraItemInput[];
}