# lib/api-zod/src/generated/types/compraItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Boutique Pro - API de gestión para boutique venezolana
 * OpenAPI spec version: 0.1.0
 */

export interface CompraItem {
  id: number;
  productoId: number;
  nombreProducto: string;
  codigoProducto?: string;
  cantidad: number;
  costoUnitarioUsd: number;
  subtotalUsd: number;
}