# lib/api-zod/src/generated/types/compraItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Boutique Pro - API de gestión para boutique venezolana
 * OpenAPI spec version: 0.1.0
 */

export interface CompraItemInput {
  productoId: number;
  cantidad: number;
  costoUnitarioUsd: number;
}