# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Boutique Pro - API de gestión para boutique venezolana
 * OpenAPI spec version: 0.1.0
 */

export interface Dashboard {
  ventasHoy: number;
  ventasHistoricas: number;
  valorInventario: number;
  totalProductos: number;
  productosStockBajo: number;
  facturasHoy: number;
  gananciasDia: number;
  gananciasHistoricas: number;
  totalComprasHoy: number;
}