# lib/api-zod/src/generated/types/producto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Boutique Pro - API de gestión para boutique venezolana
 * OpenAPI spec version: 0.1.0
 */

export interface Producto {
  id: number;
  codigo: string;
  nombre: string;
  categoria: string;
  talla: string;
  color: string;
  precioUsd: number;
  costoUsd: number;
  stock: number;
  stockMinimo: number;
  creadoEn: string;
}